refactor(commit-review): move the findings sort order into utils
sortFindingsForReview lived in composables/useCommitReviewNav.ts, but
utils/reviewFixPrompt.ts (a pure utils module) needed the same order and
was importing it from there, inverting the established utils-do-not-
depend-on-composables direction (Task 0's unifiedDiff.ts/editableTarget.ts
precedent). Moves it to utils/reviewFindingsSort.ts; useCommitReviewNav.ts
re-exports it verbatim for back-compat with CommitReviewModal.vue's import.

disabled
